# Salsa20 cleanups/performance optimizations.