The performance optimizations come mostly from processing a block at a time in the common case (same as my 'chacha20_perf' branch).

The cleanups aim to increase readability/maintainability of the code (storing the state as an array of 'u32's is more natural and how the papers are written), and also results in a fairly minor performance increase.

Before: salsa20::bench::salsa20_64k 403337 ns/iter (+/- 3252) = 162 MB/s
After: salsa20::bench::salsa20_64k 288177 ns/iter (+/- 1947) = 227 MB/s