@Yawning 2015-01-28
Ok, I'm done for real now.  I was going to stop after doing the cleanup pass, but decided I might as well SIMD the thing.  I'm fairly happy with where the performance is at, and I don't think I compromised readability too much.

After (again): salsa20::bench::salsa20_64k  242093 ns/iter (+/- 1784) = 270 MB/s