@DaGenix 2015-01-30
I had the one minor comment, but, otherwise, this looks good to me. A nice performance improvement is always exciting.

@andreasots - do you have any thoughts?