@Yawning 2015-01-30
@DaGenix - Indeed.  I fixed the comment.  Minor brainfart since I've been poking at both algorithms.