@andreasots 2015-01-30
Aren't `quarterround_a` and `quarterround_b` essentially `columnround` and `rowround`? Other than that it looks good.