@Yawning 2015-01-30
@andreasots Fixed.  I went and also renamed the swizzle_foo macros for more clarity.